backend/requirements.txt
# FounderOS Backend Dependencies
# The BEST AI Platform for Founders

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Database
supabase>=2.3.0

# AI/LLM
google-generativeai>=0.3.2
langchain>=0.1.0
langchain-core>=0.1.10
langchain-community>=0.0.10
langchain-google-genai>=0.0.6
langgraph>=0.0.26

# HTTP Client
httpx>=0.25.0
aiohttp>=3.9.0

# Environment
python-dotenv>=1.0.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
python-multipart>=0.0.6

# WebSocket support
websockets>=12.0

# Scheduler (Predis.ai-style auto-posting)
apscheduler>=3.10.0

# Social Media APIs
requests>=2.31.0
